More replacements

+++ compressed_sigma/src/amortized_homomorphism.rs
@@ -21,7 +21,7 @@ use rayon::prelude::*;
 #[derive(Clone, Debug, PartialEq, Eq, CanonicalSerialize, CanonicalDeserialize)]
 pub struct RandomCommitment<G: AffineRepr> {
     /// Maximum size of the witness vectors
-    pub max_size: u64,
+    pub max_size: u32,
     /// Random vector from Z_q^n
     pub r: Vec<G::ScalarField>,
     /// A = \vec{g}^{\vec{r}}
@@ -43,15 +43,15 @@ where
     pub fn new<R: RngCore, F: Homomorphism<G::ScalarField, Output = G>>(
         rng: &mut R,
         g: &[G],
-        max_size: usize,
+        max_size: u32,
         f: &F,
         blindings: Option<Vec<G::ScalarField>>,
     ) -> Result<Self, CompSigmaError> {
-        if g.len() < max_size {
+        if g.len() < max_size as usize {
             return Err(CompSigmaError::VectorTooShort);
         };
         let r = if let Some(blindings) = blindings {
-            if blindings.len() != max_size {
+            if blindings.len() != max_size as usize {
                 return Err(CompSigmaError::VectorLenMismatch);
             }
             blindings
@@ -61,7 +61,7 @@ where
         let t = f.eval(&r).unwrap();
         let A = G::Group::msm_unchecked(g, &r);
         Ok(Self {
-            max_size: max_size as u64,
+            max_size: max_size,
             r,
             A: A.into_affine(),
             t,
@@ -78,12 +78,7 @@ where
         let challenge_powers = get_n_powers(challenge.clone(), count_commitments);
 
         // z_tilde_i = r_i + \sum_{j in count_commitments}(witnesses_j_i * challenge^j)
-        let z_tilde = amortized_response(
-            self.max_size as usize,
-            &challenge_powers,
-            &self.r,
-            witnesses,
-        );
+        let z_tilde = amortized_response(self.max_size, &challenge_powers, &self.r, witnesses);
         Response { z_tilde }
     }
 }
@@ -95,7 +90,7 @@ where
     pub fn is_valid<F: Homomorphism<G::ScalarField, Output = G>>(
         &self,
         g: &[G],
-        max_size: usize,
+        max_size: u32,
         P: &[G],
         y: &[G],
         f: &F,
@@ -103,13 +98,13 @@ where
         t: &G,
         challenge: &G::ScalarField,
     ) -> Result<(), CompSigmaError> {
-        if g.len() < max_size {
+        if g.len() < max_size as usize {
             return Err(CompSigmaError::VectorTooShort);
         }
         if P.len() != y.len() {
             return Err(CompSigmaError::VectorLenMismatch);
         }
-        if self.z_tilde.len() != max_size {
+        if self.z_tilde.len() != max_size as usize {
             return Err(CompSigmaError::VectorLenMismatch);
         }
 
@@ -212,7 +207,7 @@ mod tests {
 
     #[test]
     fn amortization() {
-        fn check(max_size: usize) {
+        fn check(max_size: u32) {
             let mut rng = StdRng::seed_from_u64(0u64);
             let homomorphism = TestHom {
                 constants: (0..max_size)
@@ -245,10 +240,10 @@ mod tests {
 
             let rand_comm =
                 RandomCommitment::new(&mut rng, &g, max_size, &homomorphism, None).unwrap();
-            assert_eq!(rand_comm.r.len(), max_size);
+            assert_eq!(rand_comm.r.len(), max_size as usize);
             let challenge = Fr::rand(&mut rng);
             let response = rand_comm.response(vec![&x1, &x2, &x3], &challenge);
-            assert_eq!(response.z_tilde.len(), max_size);
+            assert_eq!(response.z_tilde.len(), max_size as usize);
             response
                 .is_valid(
                     &g,
@@ -307,10 +302,10 @@ mod tests {
         let evals = [eval1, eval2, eval3];
 
         let rand_comm = RandomCommitment::new(&mut rng, &g, max_size, &homomorphism, None).unwrap();
-        assert_eq!(rand_comm.r.len(), max_size);
+        assert_eq!(rand_comm.r.len(), max_size as usize);
         let challenge = Fr::rand(&mut rng);
         let response = rand_comm.response(vec![&x1, &x2, &x3], &challenge);
-        assert_eq!(response.z_tilde.len(), max_size);
+        assert_eq!(response.z_tilde.len(), max_size as usize);
 
         let start = Instant::now();
         response
